requirements.txt
# Django and Web Framework
Django>=4.2.0
djangorestframework>=3.14.0

# LeRobot - Main robotics library
lerobot

# Robot Communication
pyserial>=3.5

# Computer Vision and Camera Support
opencv-python>=4.8.0
pyrealsense2>=2.55.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Data Processing
pandas>=2.0.0
pillow>=10.0.0

# Machine Learning (for training)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Hugging Face
huggingface-hub>=0.16.0
datasets>=2.14.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
